requirements/base.txt
djangorestframework-simplejwt==5.4.0
djangorestframework==3.14.0
Django==4.2.19
pytest==8.3.4
pytest-django==4.10.0
requests==2.31.0
ruff==0.11.0